requirements.txt
distutils2-py3
beautifulsoup4
pyzbar
pillow
aiohttp
cowpy
emoji
gTTS-token
gTTS
gitpython
google-api-python-client
google_auth_oauthlib
google_images_download
googletrans
gsearch
hachoir
pybase64
python-dotenv
pytube
pytz
requests
speedtest-cli
telegraph
telethon
wikipedia
redis
dnspython
pyqrcode
pypng
qrtools
youtube-dl
pygments
matplotlib
wordcloud
lxml
pyyaml